Ran scalafmt on all added/touched files
